refactor(routes): extract playback close routes
